pyproject.toml
[build-system]
requires = ["setuptools>=61.0","wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "authSnk"
version = "0.1.0"
description = "Client Python REST para centralizar autenticação das personalizações do Sankhya"
readme = "README.md"
requires-python = ">=3.10"

authors = [
  {name = "Rodrigo Vieira", email = "rodrigo.vieira@grupoeisen.com.br"}
]

license = { text = "MIT" }

dependencies = [
    "cryptography==46.0.5",
    "fastapi==0.135.1",
    "pydantic==2.12.5",
    "python-dotenv==1.2.2",
    "Requests==2.32.5",
    "SQLAlchemy==2.0.34",
    "uvicorn==0.41.0",
    "asyncpg==0.30.0",
    "greenlet==3.2.4"
]

[project.scripts]
start = "authSnk.main:main"

[project.urls]
"Homepage" = "https://github.com/tieisen/authSnk"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
